+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add whereProp to filter for properties
+
 ## [3.0.13] - 2026-03-03
 
 ### Changed
@@ -97,6 +103,7 @@
 - Add `DeepCopy(json)`
 - Add deepCopy, deepEquals and isJson and json
 
+[Unreleased]: https://github.com/inlavigo/gg_json/compare/3.0.13...HEAD
 [3.0.13]: https://github.com/inlavigo/gg_json/compare/3.0.12...3.0.13
 [3.0.12]: https://github.com/inlavigo/gg_json/compare/3.0.11...3.0.12
 [3.0.11]: https://github.com/inlavigo/gg_json/compare/3.0.10...3.0.11